+++ robotframework_reportportal/__init__.py
@@ -1,5 +1,3 @@
-"""This package contains modules for reporting to Report Portal."""
-
 #  Copyright (c) 2023 EPAM Systems
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -12,3 +10,5 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License
+
+"""This package contains modules for reporting to Report Portal."""