+++ robotframework_reportportal/listener.py
@@ -1,5 +1,3 @@
-"""This module includes Robot Framework listener interfaces."""
-
 #  Copyright 2023 EPAM Systems
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,6 +11,8 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+
+"""This module includes Robot Framework listener interfaces."""
 
 import logging
 import os